+++ packages/@uppy/react/src/useUppy.js
@@ -18,8 +18,9 @@ export default function useUppy (factory) {
   useEffect(() => {
     return () => {
       uppy.current.close({ reason: 'unmount' })
+      uppy.current = undefined
     }
-  }, [])
+  }, [uppy])
 
   return uppy.current
 }