+++ packages/@uppy/url/src/Url.jsx
@@ -48,7 +48,8 @@ function checkIfCorrectURL (url) {
 }
 
 function getFileNameFromUrl (url) {
-  return url.substring(url.lastIndexOf('/') + 1)
+  const { pathname } = new URL(url)
+  return pathname.substring(pathname.lastIndexOf('/') + 1)
 }
 /**
  * Url
@@ -104,7 +105,7 @@ export default class Url extends UIPlugin {
       })
   }
 
-  async addFile (protocollessUrl) {
+  async addFile (protocollessUrl, optionalMeta = undefined) {
     const url = this.addProtocolToURL(protocollessUrl)
     if (!this.checkIfCorrectURL(url)) {
       this.uppy.log(`[URL] Incorrect URL entered: ${url}`)
@@ -116,6 +117,7 @@ export default class Url extends UIPlugin {
       const meta = await this.getMeta(url)
 
       const tagFile = {
+        meta: optionalMeta,
         source: this.id,
         name: this.getFileNameFromUrl(url),
         type: meta.type,