+++ packages/@uppy/react/CHANGELOG.md
@@ -7,6 +7,13 @@ Included in: Uppy v3.0.0-beta
 
 - @uppy/react: refactor to ESM (Antoine du Hamel / #3780)
 
+## 2.2.2
+
+Released: 2022-06-07
+Included in: Uppy v2.12.0
+
+- @uppy/react: Reset uppy instance when React component is unmounted (Tomasz Pęksa / #3814)
+
 ## 2.2.1
 
 Released: 2022-05-30